actualizacion del repositorio

+++ src/main/java/com/club/club/persistence/repository/BarcoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.club.club.persistence.entities.Barco;
 
-public interface BarcoRepository extends JpaRepository<Barco, Long>{
+public interface BarcoRepository extends JpaRepository<Barco, Integer>{
 
 }
