+++ src/main/java/com/club/club/persistence/repository/SalidaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.club.club.persistence.entities.Salida;
 
-public interface SalidaRepository extends JpaRepository<Salida, Long>{
+public interface SalidaRepository extends JpaRepository<Salida, Integer>{
 
 }
