+++ src/main/java/com/club/club/persistence/repository/SocioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.club.club.persistence.entities.Socio;
 
-public interface SocioRepository extends JpaRepository<Socio, Long>{
+public interface SocioRepository extends JpaRepository<Socio, Integer>{
 
 }
